monitors/requirements.txt
pycarlo==0.8.12
PyYAML==6.0.1
